requirements.txt
asgiref==3.5.2
django==4.1
sqlparse==0.4.2
djangorestframework==3.13.1
